backend/pyproject.toml
[project]
name = "coii"
version = "0.1.0"
description = "Personal Agent Workforce — tracker-driven Agent orchestration backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "apscheduler>=3.10,<4",
    "python-dotenv>=1",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "anthropic>=0.50",
    "openai>=1.50",
]

[project.scripts]
coii = "app.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
